Consolidating template expressions in core.expr.template_expr
