+++ pyomo/core/expr/numeric_expr.py
@@ -1062,189 +1062,6 @@ class _MutableSumExpression(SumExpression):
         return self
 
 
-class GetItemExpression(ExpressionBase):
-    """
-    Expression to call :func:`__getitem__` on the base object.
-    """
-    __slots__ = ('_base',)
-    PRECEDENCE = 1
-
-    def _precedence(self):  #pragma: no cover
-        return GetItemExpression.PRECEDENCE
-
-    def __init__(self, args, base=None):
-        """Construct an expression with an operation and a set of arguments"""
-        self._args_ = args
-        self._base = base
-
-    def nargs(self):
-        return len(self._args_)
-
-    def create_node_with_local_data(self, args):
-        return self.__class__(args, self._base)
-
-    def __getstate__(self):
-        state = super(GetItemExpression, self).__getstate__()
-        for i in GetItemExpression.__slots__:
-            state[i] = getattr(self, i)
-        return state
-
-    def __getattr__(self, attr):
-        if attr.startswith('__') and attr.endswith('__'):
-            raise AttributeError()
-        return GetAttrExpression((self, attr))
-
-    def getname(self, *args, **kwds):
-        return self._base.getname(*args, **kwds)
-
-    def is_potentially_variable(self):
-        if any(arg.is_potentially_variable() for arg in self._args_
-               if arg.__class__ not in nonpyomo_leaf_types):
-            return True
-        for x in itervalues(self._base._data):
-            if hasattr(x, 'is_potentially_variable') and \
-               x.is_potentially_variable():
-                return True
-        return False
-
-    def is_fixed(self):
-        if any(self._args_):
-            for x in itervalues(self._base):
-                if hasattr(x, 'is_fixed') and not x.is_fixed():
-                    return False
-        return True
-
-    def _is_fixed(self, values):
-        for x in itervalues(self._base):
-            if not x.__class__ in nonpyomo_leaf_types and not x.is_fixed():
-                return False
-        return True
-
-    def _compute_polynomial_degree(self, result):       # TODO: coverage
-        if any(x != 0 for x in result):
-            return None
-        ans = 0
-        for x in itervalues(self._base):
-            if x.__class__ in nonpyomo_leaf_types:
-                continue
-            tmp = x.polynomial_degree()
-            if tmp is None:
-                return None
-            elif tmp > ans:
-                ans = tmp
-        return ans
-
-    def _apply_operation(self, result):                 # TODO: coverage
-        return value(self._base.__getitem__( tuple(result) ))
-
-    def _to_string(self, values, verbose, smap, compute_values):
-        values = tuple(_[1:-1] if _[0]=='(' and _[-1]==')' else _
-                       for _ in values)
-        if verbose:
-            return "getitem(%s, %s)" % (self.getname(), ', '.join(values))
-        return "%s[%s]" % (self.getname(), ','.join(values))
-
-    def resolve_template(self):                         # TODO: coverage
-        return self._base.__getitem__(tuple(value(i) for i in self._args_))
-
-
-class GetAttrExpression(ExpressionBase):
-    """
-    Expression to call :func:`__getattr__` on the base object.
-    """
-    __slots__ = ()
-    PRECEDENCE = 1
-
-    def _precedence(self):  #pragma: no cover
-        return GetAttrExpression.PRECEDENCE
-
-    def nargs(self):
-        return len(self._args_)
-
-    def __getattr__(self, attr):
-        if attr.startswith('__') and attr.endswith('__'):
-            raise AttributeError()
-        return GetAttrExpression((self, attr))
-
-    def __getitem__(self, *idx):
-        return GetItemExpression(idx, base=self)
-
-    def getname(self, *args, **kwds):
-        return 'getattr'
-
-    def _compute_polynomial_degree(self, result):       # TODO: coverage
-        if result[1] != 0:
-            return None
-        return result[0]
-
-    def _apply_operation(self, result):                 # TODO: coverage
-        return getattr(result[0], result[1])
-
-    def _to_string(self, values, verbose, smap, compute_values):
-        if verbose:
-            return "getitem(%s, %s)" % values
-        return "%s.%s" % values
-
-    def resolve_template(self):                         # TODO: coverage
-        return self._apply_operation(self._args_)
-
-
-class TemplateSumExpression(ExpressionBase):
-    """
-    Expression to represent an unexpanded sum over one or more sets.
-    """
-    __slots__ = ('_iters',)
-    PRECEDENCE = 1
-
-    def _precedence(self):  #pragma: no cover
-        return TemplateSumExpression.PRECEDENCE
-
-    def __init__(self, args, _iters):
-        """Construct an expression with an operation and a set of arguments"""
-        self._args_ = args
-        self._iters = _iters
-
-    def nargs(self):
-        return len(self._args_)
-
-    def create_node_with_local_data(self, args):
-        return self.__class__(args, self._iters)
-
-    def __getstate__(self):
-        state = super(TemplateSumExpression, self).__getstate__()
-        for i in GetItemExpression.__slots__:
-            state[i] = getattr(self, i)
-        return state
-
-    def getname(self, *args, **kwds):
-        return "SUM"
-
-    def is_potentially_variable(self):
-        if any(arg.is_potentially_variable() for arg in self._args_
-               if arg.__class__ not in nonpyomo_leaf_types):
-            return True
-        return False
-
-    def _is_fixed(self, values):
-        return all(values)
-
-    def _compute_polynomial_degree(self, result):       # TODO: coverage
-        return result[0]
-
-    def _apply_operation(self, result):                 # TODO: coverage
-        raise DeveloperError("not supported")
-
-    def _to_string(self, values, verbose, smap, compute_values):
-        ans = ''
-        for iterGroup in self._iters:
-            ans += ' for %s in %s' % (','.join(str(i) for i in iterGroup),
-                                      iterGroup[0]._set)
-        val = values[0]
-        if val[0]=='(' and val[-1]==')':
-            val = val[1:-1]
-        return "SUM(%s%s)" % (val, ans)
-
-
 class Expr_ifExpression(ExpressionBase):
     """
     A logical if-then-else expression::
